Moved test_save_table_metadata to common
